Add efficient negative sampling Word2Vec using specialized Glint
This is a Word2Vec implementation using skipgram and negative sampling.
It uses a custom Glint parameter server which supports specific
parameter server operations for efficient distributed Word2Vec
computation. The implementation is based on "Network-Efficient
Distributed Word2Vec Training System for Large Vocabularies,
Erik Ordentlich et al., CIKM '16"

+++ src/main/scala/org/apache/spark/mllib/feature/BlockingBigMatrix.scala
@@ -16,7 +16,7 @@ class BlockingBigMatrix[V: ClassTag](underlying: BigMatrix[V], maxSimultaneousRe
   require(maxSimultaneousRequests > 0, "Max simultaneous requests must be non-zero")
 
   val rows: Long = underlying.rows
-  val cols: Int = underlying.cols
+  val cols: Long = underlying.cols
 
   val semaphore = new Semaphore(maxSimultaneousRequests)
 
@@ -43,13 +43,13 @@ class BlockingBigMatrix[V: ClassTag](underlying: BigMatrix[V], maxSimultaneousRe
   }
 
   override def push(rows: Array[Long],
-                    cols: Array[Int],
+                    cols: Array[Long],
                     values: Array[V])(implicit ec: ExecutionContext): Future[Boolean] = {
     blockingRequest(() => underlying.push(rows, cols, values))
   }
 
   override def pull(rows: Array[Long],
-                    cols: Array[Int])(implicit ec: ExecutionContext): Future[Array[V]] = {
+                    cols: Array[Long])(implicit ec: ExecutionContext): Future[Array[V]] = {
     blockingRequest(() => underlying.pull(rows, cols))
   }
